+++ loaderspack/src/main/java/com/agrawalsuneet/loaderspack/loaders/CurvesLoader.kt
@@ -10,7 +10,7 @@ import android.widget.RelativeLayout
 import com.agrawalsuneet.dotsloader.utils.random
 import com.agrawalsuneet.loaderspack.R
 import com.agrawalsuneet.loaderspack.basicviews.ArcView
-import com.agrawalsuneet.loaderspack.basicviews.LoaderContract
+import com.agrawalsuneet.loaderspack.contracts.LoaderContract
 import java.util.*
 
 /**
