+++ loaderspack/src/main/java/com/agrawalsuneet/loaderspack/loaders/GaugeLoader.kt
@@ -9,7 +9,7 @@ import android.widget.RelativeLayout
 import com.agrawalsuneet.dotsloader.utils.random
 import com.agrawalsuneet.loaderspack.R
 import com.agrawalsuneet.loaderspack.basicviews.ArcView
-import com.agrawalsuneet.loaderspack.basicviews.LoaderContract
+import com.agrawalsuneet.loaderspack.contracts.LoaderContract
 import com.agrawalsuneet.loaderspack.basicviews.NeedleView
 
 class GaugeLoader : RelativeLayout, LoaderContract {
